Allow megaparsec-7/8 and relax other bounds.
Makes hpython build with various newer versions of libraries.

The only considerable changes to hpython itself are due to
changes in megaparsec. We now require at least megaparsec-7.
Unfortunately, some minor CPP magic is required to support
megaparsec-8 as well.

The interface of hpython changes slightly as well, because some
of megaparsec's error types are exposed via the interface, and
they have changed in megaparsec-7.